+++ app22.py
@@ -1,0 +1,69 @@
+from flask import Flask, render_template, url_for, request, session, redirect, flash
+from flask_pymongo import PyMongo
+import bcrypt
+import os
+
+app = Flask(__name__)
+app.config['SECRET_KEY'] = 'testing'
+
+app.config['MONGO_dbname'] = 'users'
+app.config['MONGO_URI'] = 'mongodb+srv://Game_api:sI3vG3fOUjwDltxr@game.yik52gz.mongodb.net/Hackathon'
+
+mongo = PyMongo(app)
+
+@app.route("/")
+@app.route("/main")
+def main():
+    return render_template('index.html')
+
+
+@app.route("/register", methods=['POST', 'GET'])
+def register():
+    if request.method == 'POST':
+        users = mongo.db.users
+        signup_user = users.find_one({'username': request.form['username']})
+
+        if signup_user:
+            flash(request.form['username'] + ' username is already exist')
+            return redirect(url_for('register'))
+
+        hashed = bcrypt.hashpw(request.form['password'].encode('utf-8'), bcrypt.gensalt(14))
+        users.insert({'username': request.form['username'], 'password': hashed, 'email': request.form['email']})
+        return redirect(url_for('login'))
+
+    return render_template('register.html')
+
+@app.route('/')
+def hello():
+    if 'username' in session:
+        return render_template('index.html', username=session['username'])
+
+    return render_template('index.html')
+
+@app.route('/login', methods=['GET', 'POST'])
+def login():
+    if request.method == 'POST':
+        users = mongo.db.users
+        signin_user = users.find_one({'username': request.form['username']})
+
+        if signin_user:
+            if bcrypt.hashpw(request.form['password'].encode('utf-8'), signin_user['password'].encode('utf-8')) == \
+                    signin_user['password'].encode('utf-8'):
+                session['username'] = request.form['username']
+                return redirect(url_for('index'))
+
+        flash('Username and password combination is wrong')
+        return render_template('login.html')
+
+    return render_template('login.html')
+
+
+@app.route('/logout')
+def logout():
+    session.pop('username', None)
+    return redirect(url_for('index'))
+
+
+if __name__ == "__main__":
+    port = int(os.environ.get("PORT", 5000))
+    app.run(host='0.0.0.0', port=port, debug=True)
